+++ projects/store/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@studiohyperdrive/ngx-store",
-  "version": "16.0.1",
+  "version": "16.0.2",
   "peerDependencies": {
     "@angular/common": "^16.1.0",
     "@angular/core": "^16.1.0",
