+++ projects/store/src/lib/store/interfaces/index.ts
@@ -2,3 +2,4 @@ export { BasicEntityAdapterActions, BasicEntityState } from './entity-adapter-re
 export * from './base-store';
 export * from './entity-store-assets';
 export * from './store';
+export * from './store-generator';
