+++ projects/store/src/public-api.ts
@@ -2,7 +2,15 @@
  * Public API Surface of store
  */
 export { StoreService } from './lib/store/abstracts';
-export { BaseStoreAssets, EntityStoreAssets } from './lib/store/interfaces';
+export {
+	BaseStoreAssets,
+	EntityStoreAssets,
+	NgxStore,
+	BaseStoreActions,
+	BaseStoreSelectors,
+	EntityStoreActions,
+	EntityStoreSelectors,
+} from './lib/store/interfaces';
 export {
 	createStoreAssets,
 	dispatchDataToStore,
